chore: revert version bump
