+++ tests/instant-fallback.test.js
@@ -1,4 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { createKnexMock } from './helpers/knex-mock.js';
 
 vi.mock('../src/ptosc-runner.js', () => ({
   buildPtoscArgs: vi.fn(() => []),
@@ -12,33 +13,22 @@ beforeEach(() => {
   vi.clearAllMocks();
 });
 
-function createKnexMock({ version = '8.0.0', instantError } = {}) {
-  const knex = () => ({
-    where() { return this; },
-    update: async () => 1,
-    select() { return this; },
-    first: async () => ({ is_locked: 0 }),
-  });
-  knex.schema = { hasTable: async () => true };
-  knex.client = { config: { connection: { database: 'testdb', host: 'localhost', user: 'root' } } };
-  knex.raw = vi.fn((sql) => {
-    if (sql === 'SELECT VERSION() AS version') {
-      return Promise.resolve([{ version }]);
-    }
-    if (/ALGORITHM=INSTANT/.test(sql)) {
-      if (instantError) return Promise.reject(instantError);
-      return Promise.resolve();
-    }
-    return Promise.resolve();
-  });
-  return knex;
-}
 
 describe('INSTANT fallback', () => {
   it('routes error 4092 through ptosc', async () => {
     const err = new Error('unsupported');
     err.errno = 4092;
-    const knex = createKnexMock({ instantError: err });
+    const knex = createKnexMock({
+      rawImpl: vi.fn((sql) => {
+        if (sql === 'SELECT VERSION() AS version') {
+          return Promise.resolve([{ version: '8.0.0' }]);
+        }
+        if (/ALGORITHM=INSTANT/.test(sql)) {
+          return err ? Promise.reject(err) : Promise.resolve();
+        }
+        return Promise.resolve();
+      }),
+    });
     await alterTableWithPtoscRaw(knex, 'ALTER TABLE widgets ADD COLUMN foo INT');
     expect(knex.raw).toHaveBeenCalledWith(expect.stringMatching(/ALGORITHM=INSTANT/));
     expect(runPtoscProcess).toHaveBeenCalled();
@@ -46,14 +36,34 @@ describe('INSTANT fallback', () => {
 
   it('routes unsupported INSTANT message through ptosc', async () => {
     const err = new Error('ALGORITHM=INSTANT is not supported');
-    const knex = createKnexMock({ instantError: err });
+    const knex = createKnexMock({
+      rawImpl: vi.fn((sql) => {
+        if (sql === 'SELECT VERSION() AS version') {
+          return Promise.resolve([{ version: '8.0.0' }]);
+        }
+        if (/ALGORITHM=INSTANT/.test(sql)) {
+          return Promise.reject(err);
+        }
+        return Promise.resolve();
+      }),
+    });
     await alterTableWithPtoscRaw(knex, 'ALTER TABLE widgets ADD COLUMN foo INT');
     expect(knex.raw).toHaveBeenCalledWith(expect.stringMatching(/ALGORITHM=INSTANT/));
     expect(runPtoscProcess).toHaveBeenCalled();
   });
 
   it('skips INSTANT on MySQL 5.7', async () => {
-    const knex = createKnexMock({ version: '5.7.34' });
+    const knex = createKnexMock({
+      rawImpl: vi.fn((sql) => {
+        if (sql === 'SELECT VERSION() AS version') {
+          return Promise.resolve([{ version: '5.7.34' }]);
+        }
+        if (/ALGORITHM=INSTANT/.test(sql)) {
+          return Promise.resolve();
+        }
+        return Promise.resolve();
+      }),
+    });
     await alterTableWithPtoscRaw(knex, 'ALTER TABLE widgets ADD COLUMN foo INT');
     expect(knex.raw).toHaveBeenCalledWith('SELECT VERSION() AS version');
     expect(knex.raw).not.toHaveBeenCalledWith(expect.stringMatching(/ALGORITHM=INSTANT/));
